cleanup all for resources

+++ controllers/restore_post.go
@@ -339,6 +339,10 @@ func deleteDynamicResourcesForBackup(
 	}
 	labelSelector := fmt.Sprintf("%s, %s notin (%s), %s",
 		BackupNameVeleroLabel, BackupNameVeleroLabel, backupName, genericLabel)
+	if restoreOptions.cleanupType == v1beta1.CleanupTypeAll {
+		// get all resources, including user created
+		labelSelector = genericLabel
+	}
 	if otherLabels != "" {
 		labelSelector = fmt.Sprintf("%s, %s", labelSelector, otherLabels)
 	}
@@ -372,15 +376,26 @@ func deleteDynamicResourcesForBackup(
 		}
 		if dr := restoreOptions.dynamicArgs.dyn.Resource(mapping.Resource); dr != nil {
 
-			listOptions := v1.ListOptions{LabelSelector: labelSelector}
+			var listOptions = v1.ListOptions{}
+			if labelSelector != "" {
+				listOptions = v1.ListOptions{LabelSelector: labelSelector}
+			}
 			if dynamiclist, err := dr.List(ctx, listOptions); err == nil {
 				// get all items and delete them
 				for i := range dynamiclist.Items {
+
+					item := dynamiclist.Items[i]
+					if restoreOptions.cleanupType == v1beta1.CleanupTypeAll &&
+						item.GetLabels()[BackupNameVeleroLabel] == backupName {
+						// exclude here resources with the same backup as the last restore
+						continue
+					}
+
 					deleteDynamicResource(
 						ctx,
 						mapping,
 						dr,
-						dynamiclist.Items[i],
+						item,
 						restoreOptions.deleteOptions,
 						veleroBackup.Spec.ExcludedNamespaces,
 					)
